Crate ml_rs

ml-rs

This crate implements various machine learning algorithms in Rust, built on the ndarray crate.

Currently, there is support for classification and regression models, transformation (including principal component analysis) and metrics to evaluate model performance.

Setup

This crate uses BLAS/LAPACK to accelerate the linear algebra operations needed. This crate is built on ndarray-linalg, which supports the following backends:

Netlib

OpenBLAS

Intel MKL. You can choose which library you wish to link to. Roughly speaking, you should expect Netlib to be slower than the other two choices, and if you are deploying on Intel processors, MKL is likely to perform best (though the library is not open-source).

On Debian systems, a working configuration can be obtained by installing the packages libopenblas-base, libopenblas-dev and liblapacke-dev, and adding

openblas-src = { version = "0.7", features = ["system"] }

to your Cargo.toml, which will link to the system OpenBLAS.

Quick Start Guide

Load data from CSV to a 2-dimensional array:

use ml_rs::preprocessing::CsvReader;
use ndarray::Array2;
use std::fs::File;

let csv_file = File::open("data.csv").unwrap();
let reader = CsvReader::new(&csv_file);
// Pass the number of rows and columns expected
let n_columns = 5;
let n_rows = 1000;
let data: Array2<f64> = reader.read(n_rows, n_columns);

Separate into data matrix and target vector:

use ndarray::{Axis, s};
// Choose the last column of the `data` array to be the target
let feature_col_index = 4;
let x = data.slice(s![.., 0..feature_col_index]);
let y = data.index_axis(Axis(1), feature_col_index);

Perform regression:

use ml_rs::regression::linear::LinearRegression;
use ml_rs::regression::Regressor;
let mut lm = LinearRegression::new();
// Fit a linear regression model to the data, unwrapping to check for errors
lm.fit(x, y).unwrap();
// Get the predicted values for `x` given by the regression model
let y_pred = lm.predict(x);

Measure the performance of a model:

use ml_rs::metrics::accuracy_score;
// We own `y_pred`, so we need to return a view, which means we
// don't consume it when calculating accuracy.
let train_accuracy = accuracy_score(y, y_pred.view());
println!("Training set accuracy: {}%", train_accuracy * 100.0);

Classification works very similarly to regression: for an example, see the classification module.

Modules

classification
A variety of supervised classification models to use with numeric data.

metrics
A collection of metrics to measure the performance of classification and regression models.

preprocessing
Utilities including loading data from CSV files to arrays to input into models.

regression
A collection of supervised regression models to predict continuous variables from data.

transformation
Procedures to perform scaling, dimensionality reduction and other transformations on the data before input into a model.

Enums

Error
The main error type which represents an error in a model or transformer. As there are many commonalities between classifiers, regressors and transformers, this general error is returned during the use of any of these objects.