Module ml_rs::classification

A variety of classification models. The linear module includes linear models such as LogisticRegression, which performs logistic regression on the data.

Modules

linear
Classifiers based on linear regression (which, despite its name, is a classification model).

Structs

MajorityClassifier
A classifier which learns the most common class and predicts this class for all unseen data.

TrivialClassifier
A trivial classifier that is initialised with a class label and outputs that label for any sample it is given.

Traits

Classifier
This trait represents a classifier that can be fit on numeric data and outputs a discrete prediction of the correct class.

ProbabilityBinaryClassifier
A binary classifier that can return calibrated probability estimates in the range $[0, 1]$ for a given sample.

Functions

labels_binary
Convenience function to verify whether an array of labels can be used in a binary classifier.