Module ml_rs::classification::linear

Classifiers based on linear regression (which, despite its name, is a classification model).

Structs

BFGSSolver
Solver for the LogisticRegression classifier implementing the Broyden–Fletcher–Goldfarb–Shanno (BFGS) method.

IRLSSolver
Solver for the LogisticRegression classifier implementing the iteratively reweighted least squares (IRLS) method.

LogisticRegression
A classifier implementing the logistic regression model. Logistic regression models can be used for binary classification problems and may be extended through multinomial logistic regression.

Traits

LogisticRegressionSolver
Represents a backend for the LogisticRegression classifier which calculates appropriate weights. See BFGSSolver and IRLSSolver for concrete implementations of this trait.