Struct ml_rs::classification::linear::IRLSLogisticRegression

pub struct IRLSLogisticRegression {
    pub max_iter: u64,
    // some fields omitted
}

A classifier implementing the logistic regression model fit using an iteratively reweighted least squares (IRLS) method as described by Hastie et al.

Model

For a general description of the logistic regression model, see the LogisticRegression classifier. This model uses the same principle, and is still fit using maximum likelihood estimation. Instead of solving for the maximum likelihood estimator using gradient-based methods, this classifier uses a method known as iteratively reweighted least squares.

Let $X$ be the data matrix with corresponding labels $y$, $w_i$ the $i$th approxmation for the weights, $p$ the vector with entries $p_j = p(x_j; w_i)$, $W$ the diagonal matrix with entries $W_{jj} = p(x_j; w_i)(1 - p(x_j; w_i))$, and $z = Xw_i + W^{-1} (y - p)$.

The next approximation is given by solving the weighted least squares (WLS) problem $$w_{i + 1} = \argmin_w (z - Xw)^T W (z - Xw),$$ where this more obviously a least squares problem by rewriting the expression as $$w_{i + 1} = \argmin_w \lVert W^{\frac12} (z - Xw) \rVert,$$ recalling that the traditional least squares problem is to minimise $\lVert Ax - b \rVert$ over all $x$.

Internally, we solve the least squares problem $$w_{i + 1} = \argmin_w \lVert b - Cw \rVert,$$ where $b = W^{\frac12} z$ and $C = W^{\frac12} X$ using singular value decompositions.

References

[1] Hastie et al, The Elements of Statistical Learning: Data Mining, Inference and Prediction, Springer, New York, NY, 2001, 1st ed, pp. 95–100.

Fields
max_iter: u64

The maximum number of iterations of the iteratively reweighted least squares algorithm used when fitting to the data. The default is 10.

Implementations

impl IRLSLogisticRegression

pub fn new() -> IRLSLogisticRegression

Creates a new IRLSLogisticRegression classifier without any weights, which may then be trained using Classifier::fit().