Trait ml_rs::classification::Classifier

pub trait Classifier {
    fn fit<'a>(
        &mut self, 
        x: ArrayView2<'a, f64>, 
        y: ArrayView1<'a, usize>
    ) -> Result<(), Error>;

    fn predict(&self, x: ArrayView2<'_, f64>) -> Result<Array1<usize>, Error>;
}

Represents a classifier that can be fit on numeric data and outputs a discrete prediction of the correct class.

A broad variety of models exist for solving binary classification problems. These often make distinct assumptions which should be considered carefully when applying these to a particular problem.

Models that can also provide probability estimates will implement the ProbabilityBinaryClassifier trait as well as this trait.

See the module-level documentation for a broad overview of classification with ml-rs.

Required methods

fn fit<'a>(
&mut self,
x: ArrayView2<'a, f64>,
y: ArrayView1<'a, usize>
) -> Result<(), Error>

Fits the classifier to the given data matrix x and labels y. This does not support online learning and running this on a classifier that has already been fit will lose the previously learned parameters.

Arguments

x: a view to a 2-dimensional data matrix where each row corresponds to a sample. The data matrix is unchanged and only a view, which can be obtained by the .view() method of an array, is needed.

y: a view to a 1-dimensional array containing the labels for each row of x. Labels are required to be non-negative integers, so non-integer labels will need to be transformed before use.

Returns

This method returns a result that should be checked before attempting to make predictions. If fitting was successful, Ok(()) is returned, and a corresponding Error is returned if a problem occurred.

fn predict(&self, x: ArrayView2<'_, f64>) -> Result<Array1<usize>, Error>

Makes a prediction for the sample in each row of the data matrix x, returning the corresponding labels as an Array1<usize>.

Most classifiers require fitting before use with the fit() method, and failing to do so will lead to an error being returned.

Arguments

x: a view to a 2-dimensional data matrix where each row corresponds to a sample. The data matrix is unchanged and only a view, which can be obtained by the .view() method of an array, is needed.

Returns

If successful, the Result value is Ok and contains an Array1 with each element corresponding to the prediction for the corresponding row in the data matrix. Otherwise, an Error is returned.

Examples

use ndarray::array;
use ml_rs::classification::{Classifier, MajorityClassifier};

// Use some dummy training data
let x_train = array![[1.0, 2.0], [3.0, 4.0]];
let y_train = array![0, 0];

// Create a classifier and fit, unwrapping to ensure to error occurs.
let mut classifier = MajorityClassifier::new();
classifier.fit(x_train.view(), y_train.view()).unwrap();

// Use further unseen test data
let x_test = array![[-5.0, 1.0]];
// Make a prediction on `x_test`, which has 1 row, so we expect one prediction.
let y_pred = classifier.predict(x_test.view()).unwrap();
assert_eq!(y_pred, array![0]);

Loading content...