Trait ml_rs::classification::ProbabilityBinaryClassifier

pub trait ProbabilityBinaryClassifier: Classifier {
    fn predict_probability(
        &self, 
        x: ArrayView2<'_, f64>
    ) -> Result<Array1<f64>, Error>;
}

A binary classifier that can return calibrated probability estimates in the range $[0, 1]$ for a given sample.

Any classifier that implements the ProbabilityBinaryClassifier trait must also implement the Classifier trait. The predict() method of the Classifier trait is suitable for use when you only need class predictions rather than confidence levels.

Many classifiers use some sort of decision threshold, but this trait is only applied to classifiers where you can treat the returned values as a probability estimate that has been calibrated in some sense. For example, support vector machines do not (naturally) return calibrated probabilities, whereas logistic regression classifiers do.

Examples

use ml_rs::classification::linear::LogisticRegression;
use ml_rs::classification::{Classifier, ProbabilityBinaryClassifier};
use ndarray::array;

let x = array![[1.0, 2.0], [3.0, 4.0]];
let y = array![0, 1];
let mut clf = LogisticRegression::new();
clf.fit(x.view(), y.view());
let y_prob = clf.predict_probability(x.view());

Required methods

fn predict_probability(
&self,
x: ArrayView2<'_, f64>
) -> Result<Array1<f64>, Error>

Makes an estimate of the probability that each sample in the data matrix x is in class 1 (the positive class). Values close to 1.0 indicate that the classifier is highly confident that the sample is in the positive class given the training data.

Arguments

x: a view to a 2-dimensional data matrix where each row corresponds to a sample. The data matrix is unchanged and only a view, which can be obtained by the .view() method of an array, is needed.

Returns

As noted in the trait description, the returned values are calibrated probability estimates in the interval $[0, 1]$, but you should bear in mind that this is the probability estimate given the modelling assumptions and training data. If the training data are not i.i.d. or the modelling assumptions are incorrect, you may receive poor results.

Loading content...