Module ml_rs::regression

A collection of supervised regression models to predict continuous variables from data.

Regression tasks look to construct a regressor $h \colon \mathcal{X} \to \mathcal{Y}$, where instead of $\mathcal{Y}$ being a finite set as in classification, we have $\mathcal{Y}$ being a continuum, e.g. an interval $[a, b]$.

Context

The context is largely the same as in the classification case (and reading the classification module documentation should prove helpful). The main change is the the space $\mathcal{Y}$ which is now continuous, so some classification models do not have corresponding regression models, whereas others (such as least squares regression) find applications in both classification and regression.

Models

Currently, the following regressors are supported.

Linear Regression

These models are appropriate when the target $Y$ is believed to be some linear function $f(X)$ of the feature vector $X$.

regression::linear::LinearRegression.

Modules

linear

Traits

Regressor
This trait represents a regression model, used to predict a continuous value fit from some training data. Contrast this with the Classifier trait in the classification module, which assigns a discrete class (in other words, predicts the class) of some data after being fit upon some training data.