Crate ml_rs

Modules

classification
A variety of supervised classification models to use with numeric data.

metrics
A collection of metrics to measure the performance of classification and regression models.

preprocessing
Utilities including loading data from CSV files to arrays to input into models.

regression
Models to predict continuous variables from data.

transformation
Procedures to perform scaling, dimensionality reduction and other transformations on the data before input into a model.

Enums

Error
The main error type which represents an error in a model or transformer. As there are many commonalities between classifiers, regressors and transformers, this general error is returned during the use of any of these objects.